Further cleanup and corrections
removed orignal KML script for not to implicitly
re-distribute it. Even though Rop agreed that we may
use it, it doesn't make sense to distribute it at two places
on the web where ours might become outdated.
So better pull it from the original location (link given
in our code comments)